requirements.txt
streamlit==1.32.0
pandas==2.2.1
requests==2.31.0
python-dateutil==2.8.2
holidays==0.45
pycountry==23.12.11 